create redux store
